fix: fix cycle overflow

+++ tetanes-core/src/apu.rs
@@ -87,9 +87,9 @@ pub trait ApuRegisters {
 #[must_use]
 pub struct Apu {
     pub frame_counter: FrameCounter,
-    pub master_cycle: usize,
-    pub cpu_cycle: usize,
-    pub cycle: usize,
+    pub master_cycle: u64,
+    pub cpu_cycle: u64,
+    pub cycle: u64,
     pub clock_rate: f32,
     pub region: NesRegion,
     pub pulse1: Pulse,
@@ -115,7 +115,7 @@ impl Apu {
     pub const DEFAULT_SAMPLE_RATE: f32 = 44_100.0;
     // 5 APU channels + 1 Mapper channel
     pub const MAX_CHANNEL_COUNT: usize = 6;
-    pub const CYCLE_SIZE: usize = 10_000;
+    pub const CYCLE_SIZE: u64 = 10_000;
 
     /// Create a new APU instance.
     pub fn new(region: NesRegion) -> Self {
@@ -148,12 +148,13 @@ impl Apu {
     }
 
     pub fn default_channel_outputs() -> Vec<f32> {
-        vec![0.0; Self::MAX_CHANNEL_COUNT * Self::CYCLE_SIZE]
+        vec![0.0; Self::MAX_CHANNEL_COUNT * Self::CYCLE_SIZE as usize]
     }
 
     pub fn add_mapper_output(&mut self, output: f32) {
         self.channel_outputs
-            [(self.master_cycle * Self::MAX_CHANNEL_COUNT) + Channel::Mapper as usize] = output;
+            [(self.master_cycle as usize * Self::MAX_CHANNEL_COUNT) + Channel::Mapper as usize] =
+            output;
     }
 
     /// Filter and mix audio sample based on region sampling rate.
@@ -165,7 +166,7 @@ impl Apu {
         for outputs in self
             .channel_outputs
             .chunks_exact(Self::MAX_CHANNEL_COUNT)
-            .take(self.master_cycle)
+            .take(self.master_cycle as usize)
         {
             let [pulse1, pulse2, triangle, noise, dmc, mapper] = outputs else {
                 warn!("invalid channel outputs");
@@ -241,20 +242,20 @@ impl Apu {
         }
     }
 
-    pub fn clock_lazy(&mut self) -> usize {
+    pub fn clock_lazy(&mut self) -> u64 {
         self.cpu_cycle = self.cpu_cycle.wrapping_add(1);
         self.master_cycle += 1;
         if self.master_cycle == Self::CYCLE_SIZE - 1 {
             self.clock_flush()
         } else if self.should_clock() {
-            self.clock_to(self.master_cycle as u64)
+            self.clock_to(self.master_cycle)
         } else {
             0
         }
     }
 
-    pub fn clock_flush(&mut self) -> usize {
-        let cycles = self.clock_to(self.master_cycle as u64);
+    pub fn clock_flush(&mut self) -> u64 {
+        let cycles = self.clock_to(self.master_cycle);
 
         self.process_outputs();
 
@@ -281,14 +282,14 @@ impl Apu {
         self.frame_counter.should_clock(cycles) || self.dmc.irq_pending_in(cycles)
     }
 
-    fn channel_clock_to(&mut self, channel: Channel, cycle: usize) {
-        fn clock_to<T>(instance: &mut T, cycle: usize, offset: usize, outputs: &mut [f32])
+    fn channel_clock_to(&mut self, channel: Channel, cycle: u64) {
+        fn clock_to<T>(instance: &mut T, cycle: u64, offset: usize, outputs: &mut [f32])
         where
             T: Clock + TimerCycle + Sample,
         {
             while instance.cycle() < cycle {
                 instance.clock();
-                outputs[((instance.cycle() - 1) * Apu::MAX_CHANNEL_COUNT) + offset] =
+                outputs[((instance.cycle() - 1) as usize * Apu::MAX_CHANNEL_COUNT) + offset] =
                     instance.output();
             }
         }
@@ -307,8 +308,8 @@ impl Apu {
 }
 
 impl ClockTo for Apu {
-    fn clock_to(&mut self, cycle: u64) -> usize {
-        self.master_cycle = cycle as usize;
+    fn clock_to(&mut self, cycle: u64) -> u64 {
+        self.master_cycle = cycle;
 
         let cycles = self.master_cycle - self.cycle;
         trace!(
@@ -361,7 +362,7 @@ impl Default for Apu {
 impl ApuRegisters for Apu {
     /// $4000 Pulse1, $4004 Pulse2, and $400C Noise Control.
     fn write_ctrl(&mut self, channel: Channel, val: u8) {
-        self.clock_to(self.master_cycle as u64);
+        self.clock_to(self.master_cycle);
         match channel {
             Channel::Pulse1 => {
                 trace!("APU $4000 write: ${val:02X} - CYC:{}", self.cpu_cycle);
@@ -382,7 +383,7 @@ impl ApuRegisters for Apu {
 
     /// $4001 Pulse1 and $4005 Pulse2 Sweep.
     fn write_sweep(&mut self, channel: Channel, val: u8) {
-        self.clock_to(self.master_cycle as u64);
+        self.clock_to(self.master_cycle);
         match channel {
             Channel::Pulse1 => {
                 trace!("APU $4001 write: ${val:02X} - CYC:{}", self.cpu_cycle);
@@ -398,7 +399,7 @@ impl ApuRegisters for Apu {
 
     /// $4002 Pulse1, $4006 Pulse2, $400A Triangle, $400E Noise, and $4010 DMC Timer Low Byte.
     fn write_timer_lo(&mut self, channel: Channel, val: u8) {
-        self.clock_to(self.master_cycle as u64);
+        self.clock_to(self.master_cycle);
         match channel {
             Channel::Pulse1 => {
                 trace!("APU $4002 write: ${val:02X} - CYC:{}", self.cpu_cycle);
@@ -426,7 +427,7 @@ impl ApuRegisters for Apu {
 
     /// $4003 Pulse1, $4007 Pulse2, and $400B Triangle Timer High Byte.
     fn write_timer_hi(&mut self, channel: Channel, val: u8) {
-        self.clock_to(self.master_cycle as u64);
+        self.clock_to(self.master_cycle);
         match channel {
             Channel::Pulse1 => {
                 trace!("APU $4003 write: ${val:02X} - CYC:{}", self.cpu_cycle);
@@ -449,7 +450,7 @@ impl ApuRegisters for Apu {
 
     /// $4008 Triangle Linear Counter.
     fn write_linear_counter(&mut self, val: u8) {
-        self.clock_to(self.master_cycle as u64);
+        self.clock_to(self.master_cycle);
         trace!("APU $4008 write: ${val:02X} - CYC:{}", self.cpu_cycle);
         self.triangle.write_linear_counter(val);
         self.should_clock = true;
@@ -457,7 +458,7 @@ impl ApuRegisters for Apu {
 
     /// $400F Noise and $4013 DMC Length.
     fn write_length(&mut self, channel: Channel, val: u8) {
-        self.clock_to(self.master_cycle as u64);
+        self.clock_to(self.master_cycle);
         trace!("APU $400F write: ${val:02X} - CYC:{}", self.cpu_cycle);
         match channel {
             Channel::Noise => {
@@ -471,19 +472,19 @@ impl ApuRegisters for Apu {
 
     /// $4011 DMC Output Level.
     fn write_dmc_output(&mut self, val: u8) {
-        self.clock_to(self.master_cycle as u64);
+        self.clock_to(self.master_cycle);
         trace!("APU $4011 write: ${val:02X} - CYC:{}", self.cpu_cycle);
         // Only 7-bits are used
         self.dmc.write_output(val & 0x7F);
         // $4011 applies new output right away, not on timer reload.
         let offset = Channel::Dmc as usize;
-        self.channel_outputs[(self.dmc.timer.cycle * Apu::MAX_CHANNEL_COUNT) + offset] =
+        self.channel_outputs[(self.dmc.timer.cycle as usize * Apu::MAX_CHANNEL_COUNT) + offset] =
             self.dmc.output();
     }
 
     /// $4012 DMC Sample Addr.
     fn write_dmc_addr(&mut self, val: u8) {
-        self.clock_to(self.master_cycle as u64);
+        self.clock_to(self.master_cycle);
         trace!("APU $4012 write: ${val:02X} - CYC:{}", self.cpu_cycle);
         self.dmc.write_addr(val);
     }
@@ -492,7 +493,7 @@ impl ApuRegisters for Apu {
     ///
     /// $4015   if-d nt21   DMC IRQ, frame IRQ, length counter statuses
     fn read_status(&mut self) -> u8 {
-        self.clock_to(self.master_cycle as u64);
+        self.clock_to(self.master_cycle);
         let val = self.peek_status();
         trace!("APU $4015 read: ${val:02X} - CYC:{}", self.cpu_cycle);
         if Cpu::has_irq(Irq::FRAME_COUNTER) {
@@ -537,7 +538,7 @@ impl ApuRegisters for Apu {
     ///
     /// $4015   ---d nt21   length ctr enable: DMC, noise, triangle, pulse 2, 1
     fn write_status(&mut self, val: u8) {
-        self.clock_to(self.master_cycle as u64);
+        self.clock_to(self.master_cycle);
         trace!("APU $4015 write: ${val:02X} - CYC:{}", self.cpu_cycle);
         Cpu::clear_irq(Irq::DMC);
         self.pulse1.set_enabled(val & 0x01 == 0x01);
@@ -549,7 +550,7 @@ impl ApuRegisters for Apu {
 
     /// $4017 APU Frame Counter.
     fn write_frame_counter(&mut self, val: u8) {
-        self.clock_to(self.master_cycle as u64);
+        self.clock_to(self.master_cycle);
         trace!("APU $4017 write: ${val:02X} - CYC:{}", self.cpu_cycle);
         self.frame_counter.write(val, self.cpu_cycle);
     }
@@ -562,7 +563,7 @@ impl Regional for Apu {
 
     fn set_region(&mut self, region: NesRegion) {
         if self.region != region {
-            self.clock_to(self.master_cycle as u64);
+            self.clock_to(self.master_cycle);
             self.region = region;
             self.clock_rate = Cpu::region_clock_rate(region);
             self.filter_chain = FilterChain::new(region, self.sample_rate);
