+++ tetanes-core/src/mapper/m095_dxrom.rs
@@ -93,7 +93,7 @@ impl Reset for Dxrom {
     }
 }
 impl Clock for Dxrom {
-    fn clock(&mut self) -> usize {
+    fn clock(&mut self) -> u64 {
         self.inner.clock()
     }
 }
